// task_2/starwars/src/app/swapi-service/interfaces/planet.ts
export interface Planet {
  climate: string[];
  created: Date;
  diameter: number;
  edited: Date;
  films: string[];
  gravity: number;
  name: string;
  orbital_period: number;
  population: number;
  residents: string[];
  rotation_period: number;
  surface_water: number;
  terrain: string[];
  id: number;
}